mv completions fn to autocomplete file
