+++ server/src/linter.ts
@@ -1,20 +1,15 @@
 import {
-  CompletionItem,
-  CompletionParams,
   Connection,
   Diagnostic,
   DiagnosticSeverity,
 } from 'vscode-languageserver'
 import { TextDocument } from 'vscode-languageserver-textdocument'
-import { Environment, Import, Problem, validate } from 'wollok-ts'
+import { Environment, Problem, validate } from 'wollok-ts'
 import { List } from 'wollok-ts/dist/extensions'
-import { completionsForNode } from './functionalities/autocomplete/node-completion'
-import { completeMessages } from './functionalities/autocomplete/send-completion'
 import { reportValidationMessage } from './functionalities/reporter'
 import { updateDocumentSettings } from './settings'
 import { TimeMeasurer } from './time-measurer'
 import {
-  cursorNode,
   trimIn,
 } from './utils/text-documents'
 import { isNodeURI, relativeFilePath, wollokURI } from './utils/vm/wollok'
@@ -80,25 +75,6 @@ export const validateTextDocument =
     }
   }
 
-export const completions = (environment: Environment) => (
-  params: CompletionParams,
-): CompletionItem[] => {
-  const timeMeasurer = new TimeMeasurer()
-
-  const { position, textDocument, context } = params
-  const selectionNode = cursorNode(environment, position, textDocument)
-
-  timeMeasurer.addTime(`Autocomplete - ${selectionNode?.kind}`)
-
-  const autocompleteMessages = context?.triggerCharacter === '.' && !selectionNode.parent.is(Import)
-  if (autocompleteMessages) {
-    // ignore dot
-    position.character -= 1
-  }
-  const result = autocompleteMessages ? completeMessages(environment, selectionNode) : completionsForNode(selectionNode)
-  timeMeasurer.finalReport()
-  return result
-}
 
 export const generateErrorForFile = (connection: Connection, textDocument: TextDocument): void => {
   const documentUri = wollokURI(textDocument.uri)
